Change in pw_constant indexerror sur chili

+++ ruptures/datasets/pw_constant.py
@@ -29,7 +29,7 @@ def pw_constant(n_samples, n_features=1, n_bkps=3, noisy=False, sigma=1.,
     bkps = draw_bkps(n_samples, n_bkps)
     # we create the signal
     signal = np.empty((n_samples, n_features), dtype=float)
-    tt_ = np.arange(n_samples, dtype=int)
+    tt_ = np.arange(int(n_samples))
     delta_min, delta_max = delta
     # mean value
     center = np.zeros(n_features)
